All greif related checks now respect ignoregreifdetection and ban/kick greif flags.
Refactored common ban/kick logic into helper method in tools.
Added ForceKick for when wanting to disallow players to join when sever is full or name is too long or not on whitelist etc.
Updated OnJoin to set e.Handled when players are force kicked during join.
Logging all players that successfully join the server.