Added Application tokens that can never be deleted, and are loaded either by the command line arg, or the list in the config file.  Other plugins can easily add their own tokens into the dictionary as well, however, I believe the config file approach works best for transparency to the user.
Currently, this is an extremely complex looking field in the config file now.  I don't think we should add a placeholder tho as servers wouldn't realize they had a rest token there.  Perhaps when we do a dump of the config file, someone can add the format for this field.

Closes #520