Added player reference to the ChestOpenEventArgs
The ChestOpen handler doesn't have reference to the player, which makes
it difficult to search through who is opening the chest. This change
rectifies that by passing reference of the Player to the OnChestOpen
method and then through the handler. The Player reference was already
available, but not being passed.