Add permission to ignore max HP/MP, add config option for max HP/MP, possibly fix SSC
Fixing SSC is done by sending packet 39 (item disown) to the client.
Until a response, which is packet 22, is received from the client, all
SSC-related packets are completely ignored. This ensures that the SSC
data which is sent is not dirtied due to packets which are processed
during RestoreCharacter or slightly afterwards.