Swap 0 with PlayerAnnounceResult.None
In 226994c1b286c398b639302a41cd27257073a1e2 I added a magic 0 instead of
a call to an enum because I was looking at an outdated diff that didn't
have `None` as an option. Turns out this was added a day later in a
commit I didn't see originally. Changed to `None` to make code better
documentation wise.