i18nify some commands
This replaces several strings with i18n-friendly variants so that they
can be internationalized.

Sometimes, the command in English has been reworded, because it was
dumb.

I stopped at GrabUserUserInfo, mostly because I was tired. My
rationale/thinking is that rather than doing a huge commit with the
entire file (which may take a while to build up) it would be better to
get more strings to translators faster. This is because these strings,
in particular, are some of the most important ones to translate.